refactor(rpmutils): split metadata cache helpers
